Now able to retrieve electiveSubjects for round 1
You can call the facade, and get all the courses for round 1. It will
of course return as json.

NOTE: Added dependency injection for entity manager!

IMPORTANT: So far, i haven’t been able to use Iterator or enhanced for
loop in the facade, to loop over the resultlist, so i’ve just used
standard for loop.

Commit by: Thomas